Refactor S3 connection operations to connections utility
